requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.17
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
load-dotenv==0.1.0
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.1.1
typing_extensions==4.4.0
urllib3==1.26.14
zipp==3.13.0
